Refactor auth
