+++ app/src/Controller/CoordinatesController.php
@@ -1,8 +1,7 @@
 <?php
 namespace App\Controller;
 
-use App\Controller\AppController;
-use Cake\Core\Exception\Exception;
+use Cake\Event\Event;
 
 /**
  * Coordinates Controller
@@ -11,6 +10,14 @@ use Cake\Core\Exception\Exception;
  */
 class CoordinatesController extends AppController
 {
+    public function beforeFilter(Event $event)
+    {
+        parent::beforeFilter($event);
+        $this->Auth->allow(
+            ['view']
+        );
+    }
+
     /**
      * Index method
      *
