+++ app/src/Controller/FavoritesController.php
@@ -1,16 +1,13 @@
 <?php
 namespace App\Controller;
 
-use App\Controller\AppController;
-
 /**
  * Favorites Controller
  *
  * @property \App\Model\Table\FavoritesTable $Favorites
  */
-class FavoritesController extends PostsController
+class FavoritesController extends AppController
 {
-
     /**
      * Index method
      *
@@ -108,17 +105,5 @@ class FavoritesController extends PostsController
             $this->Flash->error(__('The favorite could not be deleted. Please, try again.'));
         }
         return $this->redirect(['action' => 'index']);
-    }
-
-    public function isAuthorized($user = null)
-    {
-        $action = $this->request->params['action'];
-
-        if (in_array($action, ['add', 'index', 'delete'])) {
-            return true;
-        }
-
-        // 抜けたものはとりあえず非許可
-        return false;
     }
 }
