--- app/src/Controller/PostsController.php
@@ -1,64 +1,0 @@
-<?php
-/**
- * CakePHP(tm) : Rapid Development Framework (http://cakephp.org)
- * Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
- *
- * Licensed under The MIT License
- * For full copyright and license information, please see the LICENSE.txt
- * Redistributions of files must retain the above copyright notice.
- *
- * @copyright Copyright (c) Cake Software Foundation, Inc. (http://cakefoundation.org)
- * @link      http://cakephp.org CakePHP(tm) Project
- * @since     0.2.9
- * @license   http://www.opensource.org/licenses/mit-license.php MIT License
- */
-namespace App\Controller;
-
-use App\Controller\AppController;
-
-/**
- * Application Controller
- *
- * Add your application-wide methods in the class below, your controllers
- * will inherit them.
- *
- * @link http://book.cakephp.org/3.0/en/controllers.html#the-app-controller
- */
-class PostsController extends AppController
-{
-    public function initialize()
-    {
-        parent::initialize();
-        $this->loadComponent('Auth', [
-            'authorize' => 'Controller',
-            'authenticate' => [
-                'Form' => [
-                    'fields' => [
-                        'username' => 'mail',
-                        'password' => 'password'
-                    ]
-                ]
-            ],
-            'loginAction' => [
-                'controller' => 'Users',
-                'action' => 'login'
-            ],
-            'logoutAction' => [
-                'controller' => 'Users',
-                'action' => 'logout'
-            ],
-            'unauthorizedRedirect' => $this->referer()
-        ]);
-
-        $this->Auth->allow(['display']);
-
-        $this->Auth->loginRedirect = [];
-        $session = $this->request->session();
-        $session->write('Auth.redirect', ['controller' => 'Pages', 'action' => 'home']);
-    }
-
-    public function isAuthorized($user)
-    {
-        return false;
-    }
-}
