+++ app/src/Controller/ShopsController.php
@@ -1,7 +1,7 @@
 <?php
 namespace App\Controller;
 
-use App\Controller\AppController;
+use Cake\Event\Event;
 
 /**
  * Shops Controller
@@ -10,6 +10,13 @@ use App\Controller\AppController;
  */
 class ShopsController extends AppController
 {
+    public function beforeFilter(Event $event)
+    {
+        parent::beforeFilter($event);
+        $this->Auth->allow(
+            ['index', 'view']
+        );
+    }
 
     /**
      * Index method
